ci(Move R.1b): clippy-gate the stdlib satellite crates (-D warnings)
The feature-gated stdlib satellite crates (molt-runtime-X) are the DEFAULT
native source for their modules, but only molt-backend was clippy-gated in CI —
so lints drifted into the satellites unnoticed:

  * molt-runtime-crypto/src/hashlib.rs:964 — unnecessary `(n as u64)` cast where
    `n` is already `u64` (scrypt log_n computation). Deleted the cast.
  * molt-runtime-regex/src/regex.rs:645 — clippy::while_let_loop on the dict-iter
    consume loop. The idiomatic `while let` rewrite is NOT applied because the
    in-tree copy (builtins/regex.rs) consumes a raw-bits `molt_iter_next` and
    breaks via `as_ptr()`, so a `while let` on the satellite's Option-returning
    bridge `molt_iter_next` would diverge the two copies under the satellite-
    parity guard (residual 32 -> 33). Suppress the lint with a documented
    module-level `#![allow(clippy::while_let_loop)]` (parity-neutral: the parity
    normalizer strips inner attributes) until the Move R.2 access-layer collapse
    unifies the two `molt_iter_next` contracts.

Then add a "Clippy (runtime satellites)" CI step (sibling of the existing
molt-backend clippy gate, same guarded_exec.py wrapper) that runs
`cargo clippy --all-targets -- -D warnings` over an explicit `-p` list of all 21
satellite crates. Explicit `-p` (not `--workspace`) is used so molt-backend stays
on its own `--features native-backend` gate and no crate is double-checked.

Verified all 21 satellites are clippy-clean under `--all-targets -- -D warnings`
before wiring the gate, and the satellite-parity guard stays green (total 2662),
so the new step is green on landing.

+++ runtime/molt-runtime-crypto/src/hashlib.rs
@@ -961,7 +961,7 @@ pub extern "C" fn molt_scrypt(
                 "Invalid parameter combination for n, r, p, maxmem.",
             );
         }
-        let log_n = (64 - (n as u64).leading_zeros() - 1) as u8;
+        let log_n = (64 - n.leading_zeros() - 1) as u8;
         let params = match scrypt::Params::new(log_n, r as u32, p as u32, 32) {
             Ok(val) => val,
             Err(_) => {
